backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
rq
python-jose[cryptography]
passlib[bcrypt]
python-multipart
